Split general lock tests and default-specific ones
